requirements.txt
asgiref==3.2.3
Django==3.0.7
django-mysql==3.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
mysqlclient==1.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
